Move RealtimeSuggestionIndicatorController to its own file
